Improve DMD timings for the Sega 192x64 one, matching the real one more closely (also a slight improvement for the Sega/DE/Stern 128x32 one), based on schematics (12MHz) and IRQ timings by lucky1 (thanks!!)
also improve precision of timed_interrupts_per_second

git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@4941 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/cpuexec.h
@@ -51,7 +51,7 @@ struct MachineCPU
 	void 		(*vblank_interrupt)(void);	/* for interrupts tied to VBLANK */
 	int 		vblank_interrupts_per_frame;/* usually 1 */
 	void 		(*timed_interrupt)(void);	/* for interrupts not tied to VBLANK */
-	int 		timed_interrupts_per_second;
+	double 		timed_interrupts_per_second;
 	void *		reset_param;				/* parameter for cpu_reset */
 	const char *tag;
 };
